+++ CHANGELOG.md
@@ -2,10 +2,37 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [38.4.0] 2025-11-26
+## [38.8.0] 2025-12-15
 ### Added
 - New template option allowExternalURL to allow external href links in
   rendering-xhtml extension.
+
+## [38.7.2] 2025-12-15
+### Fixed
+- Support both ordinalValue and itemWeight extensions on ValueSet.expansion
+  for calculatedExpressions.
+
+## [38.7.1] 2025-12-08
+### Added
+- Show an error message on the question if invalid HTML tags are found
+  on answerOptions.
+
+## [38.7.0] 2025-12-05
+### Added
+- Support for template-based extraction.
+
+## [38.6.0] 2025-12-05
+### Added
+- Fully support minOccurs and maxOccurs extensions.
+
+## [38.5.0] 2025-11-24
+### Added
+- New template option readonlyMode to render the Questionnaire with
+  all the input fields disabled.
+
+## [38.4.0] 2025-11-21
+### Changed
+- Updated angular and antd to v20.
 
 ## [38.3.0] 2025-09-12
 ### Added